refactor(desktop): extract shared rust util module for time + uuid
uuid_v4(), iso_now(), days_in_month(), is_leap_year(), and
epoch_secs_to_datetime() were copy-pasted across six command modules
(budget, parallel_groups, permissions, skills, workflows, config_export).
Beyond the line-count cost, the copies had already drifted: budget.rs used
unreachable!() for invalid months while config_export.rs returned 30,
silently producing the wrong day count for corrupt input.

Centralise the helpers in src/util.rs and have each module import them.
config_export's iso_to_ms now also rejects out-of-range months at parse
time instead of relying on days_in_month's fallback.

While here, hoist the three regexes in worktree::sanitize_slug into
OnceLock-backed statics so we no longer re-compile them on every call.

Net: -454 lines, single source of truth for hand-rolled calendar math.
