+++ apps/desktop/src-tauri/src/parallel_groups.rs
@@ -3,6 +3,7 @@ use tauri::{AppHandle, State};
 
 use crate::db::{Db, DbError};
 use crate::turn::{spawn_one, SpawnOneArgs, TurnError, TurnRegistry};
+use crate::util::{days_in_month, is_leap_year, ms_col_to_iso, uuid_v4};
 
 // ---------------------------------------------------------------------------
 // Structs
@@ -136,91 +137,12 @@ fn datetime_to_epoch_secs(year: i64, month: u32, day: u32, hour: u32, min: u32, 
     days * 86400 + (hour as i64) * 3600 + (min as i64) * 60 + (sec as i64)
 }
 
-/// Convert unix epoch ms → ISO 8601 string (UTC).
-fn epoch_ms_to_iso(ms: i64) -> String {
-    let secs = ms / 1000;
-    let (year, month, day, hour, min, sec) = epoch_secs_to_datetime(secs);
-    format!(
-        "{:04}-{:02}-{:02}T{:02}:{:02}:{:02}Z",
-        year, month, day, hour, min, sec
-    )
-}
-
 fn now_epoch_ms() -> i64 {
     use std::time::{SystemTime, UNIX_EPOCH};
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap_or_default()
         .as_millis() as i64
-}
-
-fn uuid_v4() -> String {
-    use sha2::{Digest, Sha256};
-    use std::time::{SystemTime, UNIX_EPOCH};
-    let t = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default();
-    let pid = std::process::id();
-    static COUNTER: std::sync::atomic::AtomicU64 = std::sync::atomic::AtomicU64::new(0);
-    let seq = COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    let input = format!("{}-{}-{}", t.as_nanos(), pid, seq);
-    let hash = Sha256::digest(input.as_bytes());
-    format!(
-        "{:02x}{:02x}{:02x}{:02x}-{:02x}{:02x}-4{:01x}{:02x}-{:02x}{:02x}-{:02x}{:02x}{:02x}{:02x}{:02x}{:02x}",
-        hash[0], hash[1], hash[2], hash[3],
-        hash[4], hash[5],
-        hash[6] & 0x0f, hash[7],
-        (hash[8] & 0x3f) | 0x80, hash[9],
-        hash[10], hash[11], hash[12], hash[13], hash[14], hash[15],
-    )
-}
-
-fn is_leap_year(y: i64) -> bool {
-    (y % 4 == 0 && y % 100 != 0) || y % 400 == 0
-}
-
-fn days_in_month(y: i64, m: u32) -> i64 {
-    match m {
-        1 | 3 | 5 | 7 | 8 | 10 | 12 => 31,
-        4 | 6 | 9 | 11 => 30,
-        2 => {
-            if is_leap_year(y) {
-                29
-            } else {
-                28
-            }
-        }
-        _ => unreachable!(),
-    }
-}
-
-fn epoch_secs_to_datetime(mut s: i64) -> (i64, u32, u32, u32, u32, u32) {
-    let sec = (s % 60) as u32;
-    s /= 60;
-    let min = (s % 60) as u32;
-    s /= 60;
-    let hour = (s % 24) as u32;
-    s /= 24;
-    let mut year: i64 = 1970;
-    loop {
-        let days = if is_leap_year(year) { 366 } else { 365 };
-        if s < days {
-            break;
-        }
-        s -= days;
-        year += 1;
-    }
-    let mut month: u32 = 1;
-    loop {
-        let d = days_in_month(year, month);
-        if s < d {
-            break;
-        }
-        s -= d;
-        month += 1;
-    }
-    let day = s as u32 + 1;
-    (year, month, day, hour, min, sec)
 }
 
 fn row_to_group(row: &rusqlite::Row<'_>) -> Result<ParallelGroupRow, rusqlite::Error> {
@@ -235,8 +157,8 @@ fn row_to_group(row: &rusqlite::Row<'_>) -> Result<ParallelGroupRow, rusqlite::E
         session_id,
         ordinal,
         merge_strategy,
-        created_at: epoch_ms_to_iso(created_at_ms),
-        completed_at: completed_at_ms.map(epoch_ms_to_iso),
+        created_at: ms_col_to_iso(created_at_ms),
+        completed_at: completed_at_ms.map(ms_col_to_iso),
     })
 }
 
@@ -275,7 +197,7 @@ pub fn parallel_group_create(
         session_id: input.session_id,
         ordinal: input.ordinal,
         merge_strategy: input.merge_strategy,
-        created_at: epoch_ms_to_iso(created_at_ms),
+        created_at: ms_col_to_iso(created_at_ms),
         completed_at: None,
     })
 }
